+++ src/app/(auth)/sign-up/_components/employee/SecurityFields.tsx
@@ -1,0 +1,43 @@
+import React from "react";
+import { UseFormReturn } from "react-hook-form";
+import {
+  FormField,
+  FormItem,
+  FormLabel,
+  FormControl,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+
+export const SecurityFields = ({ form }: { form: UseFormReturn<any> }) => {
+  return (
+    <div className="flex gap-4">
+      <FormField
+        control={form.control}
+        name="password"
+        render={({ field }) => (
+          <FormItem className="w-full">
+            <FormLabel>Нууц үг</FormLabel>
+            <FormControl>
+              <Input type="password" placeholder="********" {...field} />
+            </FormControl>
+            <FormMessage />
+          </FormItem>
+        )}
+      />
+      <FormField
+        control={form.control}
+        name="confirmPassword"
+        render={({ field }) => (
+          <FormItem className="w-full">
+            <FormLabel>Нууц үг давтах</FormLabel>
+            <FormControl>
+              <Input type="password" placeholder="********" {...field} />
+            </FormControl>
+            <FormMessage />
+          </FormItem>
+        )}
+      />
+    </div>
+  );
+};
